feat: applications channel
